+++ examples/cra-ts-kitchen-sink/.storybook/main.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 const path = require('path');
 
 module.exports = {
@@ -9,7 +10,7 @@ module.exports = {
         tsDocgenLoaderOptions: {
           tsconfigPath: path.resolve(__dirname, '../tsconfig.json'),
           shouldExtractLiteralValuesFromEnum: true,
-          propFilter: prop => {
+          propFilter: (prop: any) => {
             // Currently not working, prop.parent is always null.
             if (prop.parent) {
               return !prop.parent.fileName.includes('node_modules/@types/react/');